Move collector to package
